chore: final polish — clippy clean, log review, version in startup
- cargo clippy passes with zero warnings: collapsed ifs, redundant
  closures, manual range-contains, a type alias for the winevent
  handler cell, and two doc comments that had been stitched to the
  wrong function by past insertions (handle_mouse_event,
  Workspace::remove_window)
- log level review: info stays lifecycle-only (window open/close,
  suspend/resume, config reload, display change), debug carries
  per-event detail; no per-frame logging in reflow/animation tick
- startup banner now includes the crate version

+++ src/win/api.rs
@@ -113,7 +113,7 @@ pub fn window_exe(hwnd: HWND) -> String {
 
 /// Wrapper for `GetWindowLongPtrW` with the given index.
 pub fn get_window_long_ptr(hwnd: HWND, index: i32) -> isize {
-    unsafe { GetWindowLongPtrW(hwnd, WINDOW_LONG_PTR_INDEX(index)) as isize }
+    unsafe { GetWindowLongPtrW(hwnd, WINDOW_LONG_PTR_INDEX(index)) }
 }
 
 /// True while the HWND is still a valid window.
